Refactored dependencies
