+++ ydb/core/tx/columnshard/engines/filter.h
@@ -2,7 +2,7 @@
 
 #include "defs.h"
 #include <ydb/core/formats/arrow/program.h>
-#include <ydb/core/formats/arrow/replace_key.h>
+#include <ydb/library/formats/arrow/replace_key.h>
 #include <ydb/core/tx/columnshard/common/snapshot.h>
 
 namespace NKikimr::NOlap {
