+++ ydb/core/tx/columnshard/engines/insert_table/ya.make
@@ -12,7 +12,7 @@ SRCS(
 
 PEERDIR(
     contrib/libs/apache/arrow
-    ydb/core/formats/arrow/modifier
+    ydb/library/formats/arrow/modifier
     ydb/core/protos
     ydb/core/formats/arrow
     ydb/core/tablet_flat
