+++ ydb/core/tx/columnshard/engines/predicate/container.h
@@ -2,7 +2,7 @@
 #include "predicate.h"
 
 #include <ydb/core/formats/arrow/arrow_filter.h>
-#include <ydb/core/formats/arrow/replace_key.h>
+#include <ydb/library/formats/arrow/replace_key.h>
 
 #include <ydb/library/accessor/accessor.h>
