+++ ydb/core/tx/columnshard/engines/reader/plain_reader/iterator/fetched_data.cpp
@@ -1,8 +1,8 @@
 #include "fetched_data.h"
 
 #include <ydb/core/formats/arrow/accessor/plain/accessor.h>
-#include <ydb/core/formats/arrow/common/validation.h>
-#include <ydb/core/formats/arrow/simple_arrays_cache.h>
+#include <ydb/library/formats/arrow/common/validation.h>
+#include <ydb/library/formats/arrow/simple_arrays_cache.h>
 
 namespace NKikimr::NOlap {
