+++ ydb/core/tx/columnshard/engines/reader/plain_reader/iterator/source.cpp
@@ -4,7 +4,7 @@
 #include "plain_read_data.h"
 #include "source.h"
 
-#include <ydb/core/formats/arrow/simple_arrays_cache.h>
+#include <ydb/library/formats/arrow/simple_arrays_cache.h>
 #include <ydb/core/tx/columnshard/blobs_reader/actor.h>
 #include <ydb/core/tx/columnshard/blobs_reader/events.h>
 #include <ydb/core/tx/columnshard/hooks/abstract/abstract.h>
