+++ ydb/core/tx/columnshard/engines/storage/indexes/count_min_sketch/checker.cpp
@@ -1,6 +1,6 @@
 #include "checker.h"
 #include <ydb/core/formats/arrow/serializer/abstract.h>
-#include <ydb/core/formats/arrow/common/validation.h>
+#include <ydb/library/formats/arrow/common/validation.h>
 #include <contrib/libs/apache/arrow/cpp/src/arrow/array/array_primitive.h>
 #include <contrib/libs/apache/arrow/cpp/src/arrow/record_batch.h>
