+++ ydb/core/tx/columnshard/engines/storage/optimizer/ut/ut_optimizer.cpp
@@ -3,8 +3,8 @@
 #include <ydb/core/tx/columnshard/counters/indexation.h>
 #include <ydb/core/tx/columnshard/engines/storage/optimizer/intervals/optimizer.h>
 #include <ydb/core/formats/arrow/serializer/batch_only.h>
-#include <ydb/core/formats/arrow/simple_builder/batch.h>
-#include <ydb/core/formats/arrow/simple_builder/filler.h>
+#include <ydb/library/formats/arrow/simple_builder/batch.h>
+#include <ydb/library/formats/arrow/simple_builder/filler.h>
 #include <ydb/core/formats/arrow/serializer/full.h>
 #include <contrib/libs/apache/arrow/cpp/src/arrow/type.h>
