+++ ydb/core/tx/columnshard/engines/ut/helper.cpp
@@ -1,5 +1,5 @@
 #include "helper.h"
-#include <ydb/core/formats/arrow/simple_arrays_cache.h>
+#include <ydb/library/formats/arrow/simple_arrays_cache.h>
 
 namespace  NKikimr::NOlap::NEngines::NTest {
