+++ ydb/core/tx/columnshard/ut_rw/ut_normalizer.cpp
@@ -7,9 +7,9 @@
 
 #include <ydb/core/tx/columnshard/operations/write_data.h>
 
-#include <ydb/core/formats/arrow/simple_builder/filler.h>
-#include <ydb/core/formats/arrow/simple_builder/array.h>
-#include <ydb/core/formats/arrow/simple_builder/batch.h>
+#include <ydb/library/formats/arrow/simple_builder/filler.h>
+#include <ydb/library/formats/arrow/simple_builder/array.h>
+#include <ydb/library/formats/arrow/simple_builder/batch.h>
 
 
 namespace NKikimr {
